implement annotations overlay
This allows reading modified annotations from inside the extensions.
To achieve that, the `AnnotationsTransformation` holder is accompanied
by an `AnnotationsOverlay`, which holds an `AnnotationSet` for each
element. This set is initially identical to what's in Jandex
(initialized on first access), but can later grow or shrink.

When extension processing is complete, the annotation overlay is
invalidated, so that it doesn't unnecessarily occupy memory, and
the transformation holder is frozen, so that adding transformations
is no longer possible.