second attempt at a push-based extension API
The main difference compared to the first attempt is that
annotation transformation callbacks are either declared
or registered programmatically.

The `@Enhancement` methods are required to declare exactly
one parameter of a type from the following set:

- `ClassConfig`, `MethodConfig` or `FieldConfig`: the method
   itself is an annotation transformation callback;
- `AppArchiveConfg`: the method is executed to register
  annotation transformation callbacks programmatically.

The `@ExactType` and `@SubtypesOf` annotations can now be placed
on methods and are used in the declarative approach to constraint
for which types (and with which annotations) is the method called.

The `@Enhancement` methods can still declare parameters
of "utility" types, if that proves useful.