prototype implementation of the AppArchiveBuilder part of @Discovery
This commit also changes the implementations of extensions lookup.
Since the `@Discovery` phase must run before bean archive exists,
we can no longer search for extensions in the bean archive.

We planned to use `ServiceLoader` for this, and now is the time
to actually implement it. The rest of the extension infrastructure
is still on Jandex, which leads to some really silly code. That
should later be refactored.